Revert "Refactoring student registration"
