+++ src/main/java/com/devhamzat/student_management_system/dto/StudentDTO.java
@@ -1,4 +1,4 @@
-package com.devhamzat.student_management_system.dto.studentdto;
+package com.devhamzat.student_management_system.dto;
 
 import com.devhamzat.student_management_system.entity.Registration;
 import com.devhamzat.student_management_system.entity.Student;
@@ -23,33 +23,57 @@ public class StudentDTO {
 
     private Set<Registration> registrations;
 
+    public String getFirstName() {
+        return firstName;
+    }
 
     public void setFirstName(Student student) {
         this.firstName = student.getFirstName();
     }
 
+    public String getLastName() {
+        return lastName;
+    }
 
     public void setLastName(Student student) {
         this.lastName = student.getLastName();
+    }
+
+    public String getEmail() {
+        return email;
     }
 
     public void setEmail(Student student) {
         this.email = student.getEmail();
     }
 
+    public Gender getGender() {
+        return gender;
+    }
 
     public void setGender(Student student) {
         this.gender = student.getGender();
+    }
+
+    public StudentType getStudentType() {
+        return studentType;
     }
 
     public void setStudentType(Student student) {
         this.studentType = student.getStudentType();
     }
 
+    public LocalDate getDob() {
+        return dob;
+    }
+
     public void setDob(Student student) {
         this.dob = student.getDob();
     }
 
+    public Integer getAge() {
+        return age;
+    }
 
     public void setAge(Student student) {
         this.age = student.getAge();
@@ -63,6 +87,9 @@ public class StudentDTO {
         this.studentId = student.getStudentId();
     }
 
+    public Set<Registration> getRegistrations() {
+        return registrations;
+    }
 
     public void setRegistrations(Student student) {
         this.registrations = student.getRegistrations();
@@ -81,6 +108,7 @@ public class StudentDTO {
                 ", registrations=" + registrations +
                 '}';
     }
+
 
 
 }
