+++ src/main/java/com/devhamzat/student_management_system/entity/Student.java
@@ -99,7 +99,7 @@ public class Student {
     @Transient
     @JsonProperty("age")
     private Integer age;
-    @JsonProperty("student id")
+
     private String studentId;
