--- src/main/java/com/devhamzat/student_management_system/exceptions/CourseAlreadyExist.java
@@ -1,8 +1,0 @@
-package com.devhamzat.student_management_system.exceptions;
-
-public class CourseAlreadyExist extends RuntimeException{
-    public CourseAlreadyExist(String error) {
-
-        super(error);
-    }
-}
