--- src/main/java/com/devhamzat/student_management_system/exceptions/CourseDoesNotExist.java
@@ -1,7 +1,0 @@
-package com.devhamzat.student_management_system.exceptions;
-
-public class CourseDoesNotExist extends RuntimeException{
-    public CourseDoesNotExist(String error){
-        super(error);
-    }
-}
