--- src/main/java/com/devhamzat/student_management_system/utils/CourseType.java
@@ -1,8 +1,0 @@
-package com.devhamzat.student_management_system.utils;
-
-public enum CourseType {
-    FOUNDATIONAL_COURSE,
-    UNDER_GRADUATE_COURSE,
-    DLI_COURSE,
-    POST_GRADUATE_COURSE
-}
