+++ src/main/java/com/devhamzat/student_management_system/utils/idGenerator/PostGraduateStudentIDGenerator.java
@@ -1,4 +1,4 @@
-package com.devhamzat.student_management_system.utils.idGenerator.studentIDGenereator;
+package com.devhamzat.student_management_system.utils.idGenerator;
 
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
