+++ src/main/java/com/devhamzat/student_management_system/utils/idGenerator/UndergradauteStudentIDGenerator.java
@@ -1,6 +1,5 @@
-package com.devhamzat.student_management_system.utils.idGenerator.studentIDGenereator;
+package com.devhamzat.student_management_system.utils.idGenerator;
 
-import com.devhamzat.student_management_system.utils.idGenerator.studentIDGenereator.StudentIdGenerator;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
